separated navbars for mobile and desktop, added some further responsiveness
