+++ package.json
@@ -14,6 +14,7 @@
     "axios": "^1.8.4",
     "dompurify": "^3.2.4",
     "framer-motion": "^12.9.1",
+    "hamburger-react": "^2.5.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-parallax": "^3.5.2",
