+++ src/components/features/Gallery/Gallery.jsx
@@ -76,7 +76,7 @@ const Gallery = ({
         </h1>
       ) : (
         // sm:align-center sm:justify-center columns-3 2xl:columns-2 xl:columns-4 lg:columns-3 md:columns-2 sm:columns-1
-        <div className="Gallery   align-center justify-center columns-3  gap-4  p-4      rounded shadow-lg  ">
+        <div className="Gallery  align-center justify-center lg:columns-3  gap-4  p-4      rounded shadow-lg">
           {/* <div> */}
           {/* <> */}
           {/* 2xl:columns-5 xl:columns-4 lg:columns-3 md:columns-2 sm:columns-1 */}
@@ -85,7 +85,7 @@ const Gallery = ({
             <div
               //border-black border-4 h-full sm:max-w-sm md:max-w-md lg:max-w-lg xl:max-w-lx 2xl:max-w-2xl
               className="Asset 
-                rounded shadow-lg  hover:shadow-2xl   flex  justify-center align-center m-4 break-inside-auto overflow-auto  "
+                rounded shadow-lg  hover:shadow-2xl flex justify-center align-center m-4 break-inside-auto overflow-auto   "
               key={asset.id}
               onClick={() => handleClick(asset)}
             >
