+++ src/components/layout/Header.jsx
@@ -2,8 +2,8 @@ import React from "react";
 
 export default function Header() {
   return (
-    <div className="  flex items-center">
-      <span className="italic text-shadow-2xs text-9xl   font-thin flex align-middle text-black">
+    <div className="flex align-center justify-center">
+      <span className=" italic text-shadow-2xl text-[16vw] flex align-center justify-center font-thin text-black ">
         Gosheh Art
       </span>
     </div>
